long-form-memory-ai/backend/requirements.txt




fastapi==0.115.8
uvicorn[standard]==0.34.0
pymongo==4.6.1
motor==3.3.2
beanie==1.25.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
langchain==0.3.17
langchain-openai==0.3.3
faiss-cpu==1.8.0
sentence-transformers==3.4.1
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.28.1
tiktoken==0.8.0
pytest==8.3.4
pytest-asyncio==0.25.3
openai==1.61.0
email-validator==2.2.0